Fix all 23 audit issues: auth, tenant, budget, policy, API key, pagination
CRITICAL fixes:
- Add AuthInterceptor enforcing X-Admin-API-Key and X-Cycles-API-Key
  headers on all endpoints per spec security requirements (#1)
- Add tenant status check (SUSPENDED/CLOSED) in API key validation (#2)

HIGH fixes:
- Tenant create now idempotent: returns 200 with existing tenant on
  duplicate instead of 409 (#3)
- Set default_commit_overage_policy to REJECT on tenant create (#8)
- Add status transition validation (ACTIVE->SUSPENDED, SUSPENDED->ACTIVE,
  *->CLOSED) with suspended_at/closed_at timestamps (#7)
- Add reservation_ttl_override and rate_limits fields to Policy model (#6)
- Make BudgetCreateRequest.tenant_id optional (derive from auth) (#4)
- BudgetLedger.tenant_id kept for internal use but populated from auth (#5)

MEDIUM fixes:
- BudgetFundingRequest.idempotency_key now optional per spec (#9)
- DEBIT failure and frozen/closed budgets return 409 per spec (#10,#11)
- budgetClosed uses BUDGET_FROZEN error code instead of BUDGET_NOT_FOUND (#12)
- Add parent_tenant_id filter to tenant list (#13)
- Add scope_prefix, unit, status filters to budget list (#14)
- Add scope_pattern, status filters to policy list (#15)
- Add status filter to API key list (#16)
- Add key_id, operation, status, from, to filters to audit log list (#17)
- Implement cursor-based pagination with next_cursor on all list endpoints (#18)

LOW fixes:
- Caps fields changed from Long to Integer per spec (#20)
- BalanceController derives tenant from auth, adds scope/unit filters (#23)

+++ cycles-admin-service/cycles-admin-service-model/src/main/java/io/runcycles/admin/model/shared/Caps.java
@@ -4,9 +4,9 @@ import lombok.*;
 import java.util.List;
 @Data @Builder @NoArgsConstructor @AllArgsConstructor @JsonInclude(JsonInclude.Include.NON_NULL)
 public class Caps {
-    @JsonProperty("max_tokens") private Long maxTokens;
-    @JsonProperty("max_steps_remaining") private Long maxStepsRemaining;
+    @JsonProperty("max_tokens") private Integer maxTokens;
+    @JsonProperty("max_steps_remaining") private Integer maxStepsRemaining;
     @JsonProperty("tool_allowlist") private List<String> toolAllowlist;
     @JsonProperty("tool_denylist") private List<String> toolDenylist;
-    @JsonProperty("cooldown_ms") private Long cooldownMs;
+    @JsonProperty("cooldown_ms") private Integer cooldownMs;
 }
